loader: use _end symbol instead of end
A long time ago, the "_end" symbol replaced the "end" symbol.  The "end"
symbol was never present on the x86_64 DragonFly platform, but it was
erroneously carried inside the i386 libc for a few years (removed with
commit 8f73bea34efe21334e560a312017af168f3d1991).

The main.c file of the loader has been carrying the old version of the
symbol, but this was masked with the standard linker script which defined
"end: as an alias of "_end" and put both symbols in the binary.

This is the only instance of "end" found in the DragonFly code base, and
now it is gone.