crtstuff: Move ELF note definitions to dedicated header
Currently only note.ABI-tag is defined in crt files for the purpose
of identifying the operating system and version that created the binary.
Soon a new type of note will be created to indicate whether crt is
responsible for calling the _init function or not.  Creating a separate
header avoids duplicate note definitions.