crtstuff: Stop calling _init/_fini methods on dyn binaries
The runtime linker has already been modified to add the ability to call
_init and _fini methods on the main executable.  Previously it would
only do this for shared libraries.

A new ELF note with value 0x20 is added to signify that the crt code
no longer calls these methods so that rtld will know that it needs
to do it.  The name of the note section has been changed from
.note.ABI-tag to .note.tag.  There was an option to put the two notes
each in their own section, or put both notes in the same section.  The
latter was selected.  Since the second note is not an ABI-tag, the
section was simply renamed.  This doesn't cause any issues.

Taken from: FreeBSD SVN 232832 (2012-03-11)