gcc41: remove /usr/lib from rpath
The gcc41 specs were unconditionally adding /usr/lib to the rpath, making
it impossible to use the -z nodefaultlibs linker flag.  RTLD will look in
/usr/lib by default as the last priority of the library search path as
long as the object wasn't built with -z nodefaultlibs.

Also the __DragonFly__ macro was bumped from 1 to 3.  It's supposed to
reflect the major version of DragonFly.