kernel/acpi: Remove the redefinition of kthread_create().
It's a relict from an old pile of #ifdefs that defined various
incarnations of kthread_create()s for various FreeBSD versions.
The #ifdefs have since been removed and this #define is what
used to be the __DragonFly__ case.

There are only 3 places where kthread_create() is used in ACPI,
so remove it because it's just confusing.

While here, fix a non-constant format string in one of the
kthread_create()s.