kernel/nwfs: The NCP check only applies if it is not the module build.
Add some dependency stuff that will cause nwfs to load ncp if it is
not loaded or in the kernel.

However ncp ultimately depends on IPX which is not available as a
module, so that's the minimum that needs compiling into the kernel
in order to use nwfs.

This should really fix the kernel build.