kernel/tmpfs: Make compiling tmpfs into the kernel optional.
If the user wants to build a kernel without tmpfs built in,
allow it to him.

Our GENERIC configs have 'options TMPFS', but even if it
isn't loaded or built into the kernel, it will be kldloaded
upon mount_tmpfs(8) via vfsload(3).

This behavior is the same with our other file systems: the
module will be autoloaded at the first mount -t <type> if
it isn't either in-kernel or already loaded.