ncurses: move USE_GETCAP, update README files
The ncurses headers were generated with additional configuration switches
documented on README.DRAGONFLY.  One outcome is the USE_GETCAP should
be in the configuration header rather than in the Makefile, so that
was updated.

New files were imported to support building the terminfo database, and
the path for that will be /usr/share/terminfo.  Accordingly, the man
page and configuration header have been updated.