Unbreak world (tic)
The tic utility should be built with the ncursesw library, but that one
is not available on any system with more than 2 days since the last
build.  On those systems, the world will break pretty quickly.

The Makefile was updated to use the available ncurses library.  After
the release of DragonFly 3.2, the makefile on the master branch can
be updated to take the wide version of the ncurses library since we
only guarantee upwards compatibility for one release.