ld/gold: Add link-time optimization plugin support
The binutils 2.22 ld.bfd and ld.gold linkers were not configured to support
link-time optimization plugins (-plugin switch) previously.  This changeset
adds -plugin support to both linkers.  The ld.bfd linker was previously
built statically, but it was necessary to change this to use the libc and
libz shared libraries because the -plugin option requires the dynamic
function dlopen().

The same update was not applied to binutils 2.21 linkers.