em: Fix "Missing Interrupt Following ICR read" errata
IMS should be set to 0 before reading ICR.  It should be noted that
once IMS is 0, ICR.INT_ASSERTED will not be set, which is not mentioned
in the datasheet.

This errata exists on 82571, 82572, 82573, 82574 and 82583.  And the fix
is needed only if the legacy interrupt is used and the interrupt line is
shared with other devices.

hw.emX.irq.unshared tunable is added to give a hint to the driver that
the legacy interrupt is not shared with other devices.