Libc: Remove i386 gcc hack concerning popcountsi2
Ref: 0cfd7e44b3676f6a43164acab0aaea46365b5c53
Ref: 4b548429247a10382073a0c227af306272160d8d

Back in Oct 2010, a "horrible hack" was added to libc in order to fix a
Firefox bug concerning a late dynamic link of a libcairo.so.2 symbol.  A
weak reference to popcountsi2 was added to i386 libc in order to force it
to be available.

Up through gcc 4.6, this hack worked because the symbol is global in the
libgcc.a library, but it's a hidden symbol starting with the gcc 4.7
version of libgcc.a.  This causes a build failure on bin/rmail while
building world with gcc 4.7.

Since Oct 2010, rtld has undergone significant changes.
With the hope that the issue is obsolete, the hack is being removed.
This allows gcc 4.7 to compile DragonFly world and kernel source.
We'll have to track firefox manually to verify the issue the hack fixed
doesn't return.