ftp:  Sync with NetBSD:
make -R work the same in ftp:// like it does for http:// (i.e. work if
there is no local file).

Don't display expected EHOSTUNREACH for all but
the last connect attempts in terse mode.

When given an URL that contains "://" but is not recognised,
print an error message.

Remove unnecessary Bk/Ek pairs from SYNOPSIS.