gcc41/gcc44: Remove a modification to exclude -Wpointer-sign from -Wall.
We had this as a local change in our compilers but the problem with this
approach is that it needs to be added to each future compiler we support.

This isn't so much a problem with the compilers we have in base but with
those from pkgsrc that we support via CCVER and compilers.conf(5).

Instead, add -Wno-pointer-sign to our Makefiles whenever -Wall comes into
play.