rc.d/netif: Require cleanvar.
Before the ipfilter removal, netif required ipfilter, which required
ipmon, which required cleanvar. In other words, cleanvar was run before
netif.

With the ipfilter requirement removed from netif, cleanvar was now
ordered at some place after netif, which caused the PID file and
control socket of wpa_supplicant(8) (which is started by netif if an
ifconfig_xxx="WPA" line is found in /etc/rc.conf) to be removed from
/var/run by cleanvar after they were created.

This caused trouble for wpa_cli(8), which needs the socket.

Found-by: Johannes Hofmann <Johannes.Hofmann@gmx.de>