gcc47: Add __stack_chk_fail_local hack from gcc44
Like gcc44 before it, gcc47 is complaining about the hidden symbol of
__stack_chk_fail_local.  This patch swildner added to gcc44 will fix the
issue, but it is a bit of a hack.  A true fix may end up touching libc,
rtld, and other bits.  The original code is supposd to conserve a PIC
register.

Note that vendor-built GCCs on FreeBSD exhibit the same behavior and as
a consequence only the patched base compiler can be used to build certain
ports.

This hack should be fixed properly as soon as convenient.