gcc4X man pages: Introduce generic MLINK handling
A few months ago, the MLINK handling for binutils was updated to make
it easier to switch the primary version.  For binutils, every manpage
has a suffix based on the binutils version and the "real" man page is
linked to the primary version.

The two base compilers had the same organization, but the manual system
was not converted when binutils system was.  When gcc47 was brought in,
the manual organization wasn't implemented correctly and the gcc47 pages
ended up getting linked to the "real" man page instead of the primary
compiler gcc44.

While fixing this, bring in the binutils MLINK handling to prepare for
switching gcc47 to the primary compiler in the future.