gcc4X: rework some makefiles
These are various tweaks to improve compatibility between legacy make and
bmake.  The biggest change was moving gnu/usr.bin dependencies outside to
Makefile.inc1 as part of _startup_lib* rather than fudge it with .PHONY
targets.  This kludge was causing some issues, as well as the double .USE
trick.