Random makefile cleanup prior to bmake import
Bmake does not like to encounter "exit 0" or "exit 1".  It fails using the
exec command.  Replace instances of "exit 1" with /usr/bin/false.  There
was one instance of "exit 0" but it was at the end the target and thus
unnecessary.

The groff/tmac makefile was improved by providing a cd command for
every installed file.

Finally, one of the test makefiles was missing a single quote in one of
the targets.