libpthread: Add sem_timedwait().
It was commented out and I'm not sure why.

It needed a bit of adjustment to properly check the nanoseconds field but
with that, it passes all eleven tests of the Open POSIX Test Suite. Other
people (thesjg, vsrinivas) have tested and used it too in the past.

While here, also add a manual page (taken from FreeBSD).

In-discussion-with: vsrinivas