m4: define yylex based on __DragonFly_version
The old lex wants it and emits errors if it's not there.
The new lex doesn't want it, and emits warnings if it is there.
On gcc44, the build breaks on an error.
To make both happy, switch it on based on recent __DragonFly_version bump.

Bump __DragonFly_version to be 100% everyone can upgrade