libm: Exclude -Woverflow from -Werror in s_csqrtl.c for gcc47/i386.
The issue is that a long double in gcc47 on i386 doesn't seem to be
able to hold LDBL_MAX. At least a warning is generated. Cause
unknown, might be a bug in gcc (or our gcc's configuration).

gcc47 on FreeBSD/i386 does not show this behavior (no warning
generated at least).

Until we have a better clue, don't turn this warning into an error,
even if -Werror is specified, so that building can go on.

FreeBSD-testing: tuxillo